[e2e-test]: Refactor the storage class creation logic for WCP

+++ tests/e2e/config_change_test.go
@@ -43,11 +43,6 @@ var _ bool = ginkgo.Describe("[csi-supervisor] config-change-test", func() {
 		defer cancel()
 	})
 
-	ginkgo.AfterEach(func() {
-		if supervisorCluster {
-			deleteResourceQuota(client, namespace)
-		}
-	})
 	/*
 		Perform Password change and check if k8s resources can be modified after the password change
 		Steps:
@@ -64,8 +59,7 @@ var _ bool = ginkgo.Describe("[csi-supervisor] config-change-test", func() {
 		ginkgo.By("Invoking password change test")
 		profileID := e2eVSphere.GetSpbmPolicyID(storagePolicyName)
 		scParameters[scParamStoragePolicyID] = profileID
-		// create resource quota
-		createResourceQuota(client, namespace, rqLimit, storagePolicyName)
+
 		// Create Storage class and PVC
 		ginkgo.By("Creating Storage Class and PVC")
 		sc, pvc, err := createPVCAndStorageClass(client, namespace, nil,
@@ -73,9 +67,11 @@ var _ bool = ginkgo.Describe("[csi-supervisor] config-change-test", func() {
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
 		defer func() {
-			ginkgo.By("Deleting Storage Class")
-			err = client.StorageV1().StorageClasses().Delete(ctx, sc.Name, *metav1.NewDeleteOptions(0))
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			if !supervisorCluster {
+				ginkgo.By("Deleting Storage Class")
+				err = client.StorageV1().StorageClasses().Delete(ctx, sc.Name, *metav1.NewDeleteOptions(0))
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			}
 		}()
 
 		ginkgo.By(fmt.Sprintf("Waiting for claim %s to be in bound phase", pvc.Name))
